+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fix status message being wiped when rearranging mod order
 * Divide mod list and info panels, allowing for resizing
 * Add settings menu
+* Add UI scale setting
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)